Introduce SnapshotTransformInputsBuildOperationType
