+++ subprojects/core/src/test/groovy/org/gradle/api/internal/tasks/SnapshotTaskInputsBuildOperationResultTest.groovy
@@ -51,14 +51,14 @@ import org.gradle.test.precondition.Requires
 import org.gradle.test.preconditions.UnitTestPreconditions
 import spock.lang.Specification
 
-import static org.gradle.api.internal.tasks.SnapshotTaskInputsBuildOperationResult.FilePropertyAttribute.FINGERPRINTING_STRATEGY_ABSOLUTE_PATH
-import static org.gradle.api.internal.tasks.SnapshotTaskInputsBuildOperationResult.FilePropertyAttribute.FINGERPRINTING_STRATEGY_CLASSPATH
-import static org.gradle.api.internal.tasks.SnapshotTaskInputsBuildOperationResult.FilePropertyAttribute.FINGERPRINTING_STRATEGY_COMPILE_CLASSPATH
-import static org.gradle.api.internal.tasks.SnapshotTaskInputsBuildOperationResult.FilePropertyAttribute.FINGERPRINTING_STRATEGY_IGNORED_PATH
-import static org.gradle.api.internal.tasks.SnapshotTaskInputsBuildOperationResult.FilePropertyAttribute.FINGERPRINTING_STRATEGY_NAME_ONLY
-import static org.gradle.api.internal.tasks.SnapshotTaskInputsBuildOperationResult.FilePropertyAttribute.FINGERPRINTING_STRATEGY_RELATIVE_PATH
-import static org.gradle.api.internal.tasks.SnapshotTaskInputsBuildOperationResult.FilePropertyAttribute.from
-import static org.gradle.api.internal.tasks.SnapshotTaskInputsBuildOperationResult.FilePropertyAttribute.fromNormalizer
+import static org.gradle.api.internal.tasks.BaseSnapshotInputsBuildOperationResult.FilePropertyAttribute.FINGERPRINTING_STRATEGY_ABSOLUTE_PATH
+import static org.gradle.api.internal.tasks.BaseSnapshotInputsBuildOperationResult.FilePropertyAttribute.FINGERPRINTING_STRATEGY_CLASSPATH
+import static org.gradle.api.internal.tasks.BaseSnapshotInputsBuildOperationResult.FilePropertyAttribute.FINGERPRINTING_STRATEGY_COMPILE_CLASSPATH
+import static org.gradle.api.internal.tasks.BaseSnapshotInputsBuildOperationResult.FilePropertyAttribute.FINGERPRINTING_STRATEGY_IGNORED_PATH
+import static org.gradle.api.internal.tasks.BaseSnapshotInputsBuildOperationResult.FilePropertyAttribute.FINGERPRINTING_STRATEGY_NAME_ONLY
+import static org.gradle.api.internal.tasks.BaseSnapshotInputsBuildOperationResult.FilePropertyAttribute.FINGERPRINTING_STRATEGY_RELATIVE_PATH
+import static org.gradle.api.internal.tasks.BaseSnapshotInputsBuildOperationResult.FilePropertyAttribute.from
+import static org.gradle.api.internal.tasks.BaseSnapshotInputsBuildOperationResult.FilePropertyAttribute.fromNormalizer
 import static org.gradle.internal.fingerprint.DirectorySensitivity.DEFAULT
 import static org.gradle.internal.fingerprint.DirectorySensitivity.IGNORE_DIRECTORIES
 import static org.gradle.internal.fingerprint.LineEndingSensitivity.NORMALIZE_LINE_ENDINGS
@@ -81,7 +81,7 @@ class SnapshotTaskInputsBuildOperationResultTest extends Specification implement
         directorySensitivity << DirectorySensitivity.values()
     }
 
-    def "can convert normalizer class into a PropertyAttribute"(FileNormalizer normalizer, SnapshotTaskInputsBuildOperationResult.FilePropertyAttribute expectedPropertyAttribute) {
+    def "can convert normalizer class into a PropertyAttribute"(FileNormalizer normalizer, BaseSnapshotInputsBuildOperationResult.FilePropertyAttribute expectedPropertyAttribute) {
         expect:
         fromNormalizer(normalizer) == expectedPropertyAttribute
